refactor(audit): split the min-size concern out, drop orphaned autotile seeds
Audit of PR #983, remediation batch 7.

engine_lifecycle.cpp crossed the hard file-size ceiling on this branch, going
from just under it to 1205 lines. Its own tail comment records three earlier
splits made on exactly this event, so this is the fourth on the same seam: the
minimum-size trio — the clamp a window reports, the re-layout its change forces,
and the resize echo that tells an effect-applied rect from a user drag — answers
"how big may this window be" rather than "is it here yet", and moves to
engine_minsize.cpp. The arrivals left behind never call into it except through
the engine's own retile scheduling.

Autotile seed teardown. A screen removed from the autotile set kept its
initial-order seed, its generation counter and its strict-order marker whenever
it had no state at the current key — which is precisely the screen that was
seeded before any of its windows arrived, since the seed probe is deliberately
non-creating. The state teardown only visits screens that HAVE a state, so those
three entries outlived the removal and a strict order was consumed on re-entry
as if fresh. The reaper is no backstop: it re-arms indefinitely while any seeded
window reads as minimized, so a removed screen could hold a live timer chain as
well. They now drop in clearScreenScheduling, beside the deferred focus request
that already dies there for the same reused-id reason.

Three contracts corrected. The autotile focus capture's comment claimed the
order and focus it ships always describe one state; they share a key, not a
membership, because the order omits genuine floats while the focus slot does
not. The behaviour is left as-is and the reason written down: the receiving
engine drops a seed it cannot place, so the effect is no restore rather than a
wrong one, and a float is not where a view belongs.

pruneDesktop's comment defended dropping a per-output entry as self-healing
while the new releaseScreenOwnership doc calls the same drop a correctness bug.
Both are right, for different reasons, and the difference is now stated: the
prune path re-pushes before the count change propagates, a mode leave pushes
nothing. The tracker's own precedence list now says outright that its global-
desktop tier is a startup seed rather than a live input, which is what makes
dropping a per-output entry unrecoverable.

The animator's release path claimed a sole caller testing for DegenerateReap.
There are two, and the newer one discards the result — deliberately, because a
reap there means the leg converged onto the rect the window already occupies and
there is nothing to replace it with. Both halves now say so.

All 404 tests pass.
